refactor: Migrate from CSV to Local Connector in DashFrame
- Replaced instances of `@dashframe/connector-csv` and `@dashframe/connector-json` with `@dashframe/connector-local` across the application.
- Updated package dependencies in `package.json` files for web and e2e applications.
- Refactored local CSV handler to utilize the new local connector and adjusted data source type accordingly.
- Removed the `@dashframe/connector-csv` package, including its associated files and tests, to streamline the codebase.
- Introduced `@dashframe/csv` as a new dependency for CSV handling.

This change enhances the modularity of the connectors and aligns with the new architecture for handling local files.
