+++ packages/connector-local/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dashframe/connector-csv",
+  "name": "@dashframe/connector-local",
   "version": "0.1.0",
   "private": true,
   "type": "module",
@@ -9,21 +9,21 @@
     "build": "tsc -p tsconfig.json",
     "clean": "rm -rf dist *.tsbuildinfo",
     "dev": "tsc -w -p tsconfig.json --preserveWatchOutput",
+    "format": "prettier --check .",
+    "format:write": "prettier --write .",
     "lint": "eslint src",
     "test": "vitest run",
-    "test:coverage": "vitest run --coverage",
     "test:watch": "vitest",
     "typecheck": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
-    "@dashframe/core": "workspace:*",
+    "@dashframe/csv": "workspace:*",
+    "@dashframe/json": "workspace:*",
     "@dashframe/engine": "workspace:*",
-    "@dashframe/engine-browser": "workspace:*",
-    "apache-arrow": "^21.1.0"
+    "@dashframe/engine-browser": "workspace:*"
   },
   "devDependencies": {
-    "@vitest/coverage-v8": "^4.0.16",
     "typescript": "^5.6.0",
-    "vitest": "^4.0.16"
+    "vitest": "^3.2.4"
   }
 }
