+++ packages/connector-local/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @dashframe/connector-local
+ *
+ * Unified local file connector for DashFrame.
+ * Handles CSV and JSON file uploads with automatic format detection.
+ */
+
+export { LocalFileConnector, localFileConnector } from "./connector";
