+++ packages/csv/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@dashframe/csv",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "clean": "rm -rf dist *.tsbuildinfo",
+    "dev": "tsc -w -p tsconfig.json --preserveWatchOutput",
+    "lint": "eslint src",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:watch": "vitest",
+    "typecheck": "tsc --noEmit -p tsconfig.json"
+  },
+  "dependencies": {
+    "@dashframe/core": "workspace:*",
+    "@dashframe/engine": "workspace:*",
+    "@dashframe/engine-browser": "workspace:*",
+    "apache-arrow": "^21.1.0"
+  },
+  "devDependencies": {
+    "@vitest/coverage-v8": "^4.0.16",
+    "typescript": "^5.6.0",
+    "vitest": "^4.0.16"
+  }
+}
