+++ packages/json/.prettierignore
