+++ packages/json/src/flatten.ts
